[backport] bzl: use old method for getting VERSION stamping (#53959)
New way of handling version stamping was written with the idea that
we'll move away from the pipeline generator. But I didn't follow-up with
that impl on the tagged release versions stamping.

So the best approach at this point is to revert to using the CI
generated `VERSION`.

Need to be backported so we can have a proper release tag and version in
the binaries.

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

CI + main-dry-run to inspect images

(cherry picked from commit fc8645c3503ce4efa8838ef7a0df421098922a75)