Embeddings: retry individually when OpenAI gives us back `null` fields (#53904)
The OpenAI embeddings API will nondeterministically return `null` for
the `embedding:` field in the bulk embeddings API response payload. This
is problematic because it causes our embeddings jobs to fail
(previously, it would corrupt them, now we error). Retrying the whole
request will sometimes help, but for some repos, there might be many
duplicated chunks, so all of them need to succeed together for the full
bulk request to succeed.

This PR retries the failed chunks individually so that we don't need to
depend on all the failing chunks to succeed in the same request. It's
pretty hacky, and has a lot of duplication since we have two independent
OpenAI clients, but refactoring that will need to be a followup.