guardrails: add observability for attribution
This adds traces, logs and metrics via the Operation abstraction. It is
my first time using it, but I believe I am using it correctly. To use
this we introduce a constructor which initializes the Operation for the
attribution service.

A follow-up commit will add dashboards to grafana.

Test Plan: go test and sg start and visiting the prometheus metrics.