app: Integrate LocalGitExternalService with settings page (#53838)
This PR updates the local repositories settings page to be able to list
repositories configured via `LocalGitExternalService` and to use
`LocalGitExternalService` when adding new repositories.

The main changes are as follows:

- The GraphQL API for getting the list of repos was changed to return
the repos already grouped by service (see `GET_LOCAL_CODE_HOSTS`). This
lets us avoid having to group them by folder on the client side.
- A new GraphQL mutation (`addLocalRepositories`) was added for adding
local repositories. The mutation accepts a list of paths and the backend
creates one or more services as necessary. This also decouples the UI
from backend specifics.

I introduced a new hook (`useLocalExternalServices`) to take care of
listing, adding and removing services and my hope/plan is to use this in
the setup wizard as well.


## Test plan

Tested adding with a hardcoded path:


https://github.com/sourcegraph/sourcegraph/assets/179026/9ef1d5fc-10a1-4d43-bcac-be2ae3d14d7b