[fix] perms jobs not being scheduled if there is not already a job
Previously, the `usersWithOldestPermsQuery` and `reposWithOldestPermsQuery`
relied solely on the permission_sync_jobs table. If there was no row for the
user or repo, this query would not return that user anymore.

This resulted in users/repos never being scheduled for a permission sync in
cases when a job was not created previously (or jobs table was truncated).

Attempted to fix the queries by including those users/repos that do not
have any record in the jobs table and putting these users/repos at the top
of the queue.