bzl: fix build tags / disable cgo for server+prom  (#54359)
As a follow-up to
https://sourcegraph.slack.com/archives/C04MYFW01NV/p1687904137730639
where @unknwon spotted a failure in the pure-docker test suite, I was
ablet to root-cause the issues down to the prom-wrapper binary being
built with `netgo` but with `cgo` enabled, which led to a weird case,
where it still tried to nslookup with cgo and segfaulted.

- 56c03276dd02f4873ec013f046d0453af456d816 fixes prom-wrapper
- 7d6f928179d39bbed16f71bfa57b5a4d3e413100 conservatively disable cgo +
enable netgo on the server, even if no issues have been observed so far.
We might want to revisit this after the release.

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

- Prom-wrapper QA:
https://buildkite.com/sourcegraph/deploy-sourcegraph-docker/builds/13756
- CI (integration tests covered the server changes)

(cherry picked from commit 988e9dd18e0a527d9661f7d5f6cadc373d350213)