App - open doc links in settings in new windows (#54501)
Some docs links in the settings section navigated away from settings. It
would be best to open these in a new window so the user can view them
without losing their place in settings. This also resolves an issue in
app where allowing it to link out to the documentation.