Add default values for mode context sizes (#54548)
This adds sensible default values for models we know might be used, so
that less configuration is required for models to work correctly.

TBD: These are the allowed upstream numbers, but it looks like we so far
hardcoded to 7k instead of 9k(?). If someone knows why, let me know.

## Test plan

Added tests.