gateway: Add endpoints to read limits (#54581)
This PR adds a basic endpoint that enumerates the configured limits and
the current rate limit data. The main purpose of it is debugging, as we
don't have direct read access to redis, but we may want to build some UI
features off of this later, too.

## Test plan

Verified things work manually.