Don't overwrite external service config changes when sync is running
This fixes an issue reported by Marc on slack where changes to the external service config disappeared after a few minutes.

When this happens:
- A sync for the external service in question is in progress
- A new config is saved while the sync is still in progress

Problem:
- SyncExternalService loads the external service on entry
- Now it's syncing for an arbitrary amount of time
- And at the end, we store LastSyncAt and NextSync at values on the external service using Upsert, using the external service loaded at entry into this function

This is writing all fields from the external service passed to Upsert, which is still holding data from minutes ago, if the sync takes a while. This will effectively revert all values that changed in the meantime.

This is fixed by explicitly only writing these two fields in the syncer.