App - make App check for new repos faster (#54627)
External services are only scanned for new repos at a set interval
currently the minimum time between checks is 1 min. This overrides that
min for App and allows it to check every 15 seconds. This allows new
repos to be added more quickly during the setup flow instead of having a
wait potentially for 1 min before they are discovered and synced.