Backend: move a bunch of enterprise packages into internal (#54639)
This moves the following packages out of the `enterprise/internal`
subdirectory:
- `codycontext`
- `cody`
- `completions`
- `compute`
- `notebooks`
- `licensing`
- `license`

I'm trying to rip the band aid off while people are still on vacation.
The goal is to get rid of `enterprise/internal` entirely, then get rid
of `enterprise/cmd`. Once everything is out of `enterprise` (and even
before), we should be free to remove a bunch of the global init stuff
that exists only because of the enterprise split.