JetBrains: Auto-grow textarea (#53594)
- Closes https://github.com/sourcegraph/sourcegraph/issues/53948

I wanted to make the textarea grow automatically, as a nice-to-have for
https://github.com/sourcegraph/sourcegraph/issues/53465.
I'm calculating the new height roughly correctly (could be improved),
even taking the word wrapping into consideration (`getLineCount()` only
counts line breaks), which should be cool, but then, for whatever
reason, the UI doesn't update immediately.
I tried a bunch of simple solutions, and even this ugly `invokeLater()`
one, but nothing seems to do the trick.

Here is a 1-min demo to explain what's happening:
https://www.loom.com/share/7b5f0acebe134cf5a02e97c708763eae

No idea for the solution.

## Test plan

Attach a Loom to demo that it's working. Test it with line breaks, but
also just long lines without breaks in it. Test if the maximum and
minimum sizes are respected.

---------

Co-authored-by: Szymon Przebierowski <szymon.przebierowski@softwaremill.com>