sveltekit-prototype: Add separator component and setup storybook
This commit adds a simple separator component to be used for resizing
panels (e.g. sidebars) and is used on file tree pages.

I also setup storybooks, following the default setup instructions. I
left everybody as default except for deleting the example stories it
came with. We can always change the structure later.
Initially I wasn't able to run `pnpm run storybook`, it was throwing an
error about not being able to import `ts-dedent`. Adding it to the root
`package.json` file seems to resolve this issue.