executors: Be able to paginate over secrets (#54647)
Executor secrets could not be paginated over. Removing the fetch policy
(`no-cache`) made it so that the `previousResult` would not be `{}`.

With the `no-cache`, `previousResult` would be empty, causing an error
to be thrown since there are no `nodes` on the result.

## Test plan

Tested manually.

(cherry picked from commit 0f450d722c1bfd7768cb7d09642b41cf52626964)