sveltekit-prototyp: Properly use kit's "streaming promises" feature (#54657)
tl;dr: A little bit of housekeeping to (eventually) make use of
streaming promises and to be less unconventional.

Context: Top-level promises returned by `load` functions are
automatically resolved and the resolved values are passed to pages. This
is not practical for big or slow-loading data. The workaround I've used
so far was to return a store from the load functions instead. However
this won't work if we ever decide to use server side rendering (can't
serialize a store).
Recently SvelteKit has added support for streaming promises:
https://kit.svelte.dev/docs/load#streaming-with-promises

To make (eventually) use of this and still get benefits from using a
store we can simply return the (nested) promise from the load function
and instantiate the store inside the component.

Note: SvelteKit doesn't merge nested objects from `load` functions. If
we did something like `return {deferred: {dataA: ...}}` in a *layout*
file and `return {deferred: {dataB: ...}}` in a `page` file, the page
wouldn't have access to `dataA` unless the page load function explicitly
"forwards" it. To avoid accidental override I decided to go the other
direction: have unique toplevel names with a `deferred` sub-property. I
guess we'll see whether or not that's a good approach.

## Test plan

TypeScript is happy