uploadstore: add method to list all keys (#54682)
This adds the method `Store.List(ctx context.Context) (iterator, err)`
to the upload store.

**Why?**
Long-term we want to surface more information about the available
embeddings to the user. Currently we rely on embeddings jobs which
provide an indirect representation of the state of the embeddings store.
As a first step we have to have a unified way (GCS, S3) to find out
which embeddings we actually have (-> this PR)

Once we can list all keys, we can, for example, create a background job
that keeps a db table up-to-date which we can use to:

- quickly filter repos by their embeddings in the site-admin UI.
- list embeddings (not just jobs) and their metadata.
  
## Test plan
- new unit tests
- manual testing
- I created a bucket on GCS and configured my local instance to use that
bucket. Then I updated the embeddings worker to list all exisiting
embeddings at the end of an embeddings job. I ran a couple of embeddings
jobs and verified that the logs made sense.