env: add MustGetBytes (#54622)
This adds `env.MustGetBytes` that parses an environment variable with
`go-humanize` and updates the `EMBEDDINGS_CACHE_SIZE` env var to use
that. This is easier to read than a big blob of numbers, and easier to
configure for admins. `go-humanize` supports numbers without a label and
interprets it as bytes, just like we did before.

(cherry picked from commit d822659702d9e0a20198c2b470f7e8fa9c2138a5)