[Backport 5.1]: remove EnterpriseDB (#54704)
Backporting this PR because it moves a lot of stuff around, and I don't
want this to make other backports more difficult.

This moves a bunch of stuff out of enterprise. Notably, the final result
is `EnterpriseDB` no longer exists. This is one of the biggest sticking
points for moving other stuff out of the `enterprise` directory, so this
should help speed up future cleanups.

## Test plan

CI. 

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->