Markprompt v0.10.0 (#54643)
This PR bumps `@markprompt/web` to the latest version (0.10.0), which
uses the new Markprompt search API endpoint. This addresses the
following:
- Improved title formatting: in some situations, the kwic function was
truncating too much text around the keyword. This is no longer the case.
- Improved ranking: the new API modifies the weight put on page title
(highest), section heading, paragraph, improving the ranking of the
results.
- Improved speed: we've updated the way content and metadata is indexed,
which boosts query performance.
- Better handling of image/figure titles, such as [the Cody
page](https://docs.sourcegraph.com/cody) which uses a figure as heading.
Markprompt now grabs the `title` property in case of a figure, or the
`alt` property in case of an image.

## Test plan

Manual testing

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->