cody-gateway/embeddings: forward 400s from OpenAI (#54689)
Right now 400s get converted to 500 because we don't wrap it in a status
error. This change adds handling so that if a client gives us a request
that OpenAI treats as bad, we blame the client instead of ourselves.

## Test plan

There's a handy (integration?) test that gives us a representative
example of a bad request.