embeddings: name embeddings by id not by name
Our current naming schema (simplified version of repo name + hash) is
not ideal, because it makes it very inefficient to map embeddings to
their repo.

With this PR we migrate embeddings to the new naming schema

```
<repo ID>.embeddingindex
```

which makes the mapping trivial.

*Why?*

We currently don't have a representation of the existing embeddings in
our database. So far we used the embeddings jobs as a proxy, which is
prone to diverge from the state of the store at some point.

Another benefit of having the state of the embeddings store represented
in the db is, that we can leverage efficient unions with the repo table
to unlock improvements for admins.