graphqlbackend: update throttled to use Ctx stores
It seems we were using a weird version of throttled which I don't think
was actually based on the v2 series correctly. This updates us to the
latest version and moves us away from the removed struct GCRARateLimiter
and the deprecated interface GCRAStore.

I came across this while attempting to update deps and getting failing
compiles.

Test Plan: we have unit tests for this, so CI.