[fix] use sane default for auth provider display name
Previously we used DisplayName from CachedInfo() function return.
But refactoring to support displayPrefix and other common fields
introduced a regression where we did not use it as a default, resulting
in empty strings for DisplayName.

This commit brings back the old behavior, where we default to DisplayName
from CachedInfo() and only overwrite it with a different value if there is
one defined and it is not an empty string.