context: Schema experiments (#54641)
Relocated any experimental/significant work to this branch. Targeting
(ultimately) #51690. Immediately stacked on #54693.

Changes from #51690:
- We no longer have a `scheme`, `manager`, `name`, `version`, field. We
only have `descriptor` and `descriptor_no_suffix` for the lookup table.
Turns out we don't need to store all of them if we skewer the lookup
table.
- The lookup table now has a nullable parent identifier. Descriptors
point to their version, versions to their name, and so on. This makes a
chain of parts that form a full SCIP symbol name in the database. This
was necessary as the old schema required us to do a large search and
find the union of symbol parts, where as searching a tree/linear chain
like this can be done more efficiently (more on that to come).
- Insertion routine now has to plan slightly differently on how it
constructs identifiers (it needs to track parents between sections now).
Instead of having a set of maps, we just embed the maps into each other.
Did some generic trickery to keep the API basically the same.
- Added an out of band migration that copies the new insertion logic.
- Updated codeintel-db queries to instead of querying from the same
table five times and finding out which pieces fit into which holes, we
now always start from the (indexed) descriptor, and follow the (limited
to n = 5) chain to reconstruct the rest of the symbol name.

## Test plan

Tested locally.