Schedule embeddings jobs will not roll back transaction on read errors (#54701)
When a set of repositories are to be scheduled for embeddings (via
graphql or a policy) an error with a single repository can cause the
entire set of embeddings jobs to be rolled back. The specific error that
occurred for a recent customer issue ([slack
discussion](https://sourcegraph.slack.com/archives/C053L1AQ0BC/p1688671109829779))
was due to one of the [repos being
empty](https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/internal/embeddings/schedule.go?L36-42)
and causing all 500 repos matched by a policy to not be enqueued for
embeddings jobs. Not only do all repos have their embeddings jobs rolled
back but additionally the logging will only report one problematic repo
before exiting the scheduling logic which makes it so that the log does
not capture all of the problematic repos in one attempt.

We should instead allow read errors (e.g. reading sql or fetching
information from gitserver) to be handled but not prevent the entire
batch of jobs to be rolled back. This will allow a policy to be
partially impactful across the repositories it is concerned with.
Especially given that an empty repo won't have any need to be enqueued
for an embeddings job.

This PR changes the error handling so that repos that cannot complete an
embeddings job will still be enqueued with a `failed` state so that site
admin pages can still display which individual repos are problematic.
When a repo is enqueued with a `failed` state we do not roll back the
transaction so that successful embeddings jobs are still created with a
`queued` state.

## Test plan

will add test case to either scheduling logic tests or job execution
logic tests (depends on implementation)

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

(cherry picked from commit 57146ce983bc6515779dc0ae2bebf89f03fd5ab3)