search: use ReposMap for ListAllIndexed
Minimal is deprecated since ReposMap has a more garbage collector
friendly representation. This switches us over for ListAllIndexed calls.
Additionally we update the return type to only have the fields that will
be set by this call. This is to prevent accidental use of an unset
field.

Test Plan: CI