goroutine/periodic: avoid double-logging errors (#54874)
Avoid double-logging errors in periodic goroutines by creating an
`observation.Context` that does nothing but logging if an operation is
not provided, instead of doing the logging manually. This allows us to
remove the `simpleHandler` layer in here entirely as well.

Previously, the generated error message is noisy in Sentry feeds when an
operation is already provided, and the operation log is superior in
general (includes trace context). This change respects a provided
operation better while retaining back-compat (fallback error logging)

## Test plan

Without `goroutine.WithOperation`:

```
[   cody-gateway] ERROR periodic.sourcesSync goroutine/periodic.go:375 operation.error {"count": 1, "elapsed": 0.040589458, "error": "failed to sync dotcom-product-subscriptions: failed to list subscriptions from dotcom: returned error 401 Unauthorized: Invalid access token."}
```

With `goroutine.WithOperation` (note the trace, different root scope):

```
[   cody-gateway] ERROR cody-gateway.sourcesSync goroutine/periodic.go:375 operation.error {"TraceId": "a5373a93619b15f9600a5f473d13aaea", "SpanId": "af8738487e92812c", "count": 1, "elapsed": 0.012453917, "error": "failed to sync dotcom-product-subscriptions: failed to list subscriptions from dotcom: returned error 401 Unauthorized: Invalid access token."}
```

In the second case now only 1 entry is generated, whereas before we had
2

---------

Co-authored-by: Erik Seliger <erikseliger@me.com>