app: prepare for a more minimal distribution & reduce red-herring errors (#54898)
Almost every user report I respond to has errors like this in their
logs, which leads to users getting confused and thinking _that_ must be
their issue (when in reality, it's some less verbose log line):

```
 Docker is unavailable                                                        |
 src-cli is unavailable                                                       |
Failed to get src-cli version
exec: \"src\": executable file not found in $PATH"
Failed to get docker version
exec: \"docker\": executable file not found in $PATH"
```

The verbosity of these errors made sense when the app was a single
`sourcegraph` binary, and we wanted it to be a full Sourcegraph instance
with precise code intel, zoekt, and more. These days, however, we
removed code search from the product and it's just the Cody app. We're
still doing the same stuff behind the scenes, and it's affecting the
user experience.

What I have done is introduced an env var `APP_FULL_SOURCEGRAPH=false`
by default, which one can check via `deploy.IsAppFullSourcegraph()` to
determine if this Cody app should 'behave as a full Sourcegraph instance
does'. By default, it is off - and we can use this to start disabling
various logic while still keeping that code functional for when we _do_
want to bring it back in the future. I have used this to disable the
red-herring errors above.

## Test plan

`sg start app` ran & didn't blow up, reading the code confirms this
shouldn't affect any other deployment types.

---------

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>