all: replace "single binary OR app" with just "app" deploy type
Originally, the intent was to use the single-binary version of Sourcegraph
in many more places. e.g. inside each enterprise Docker container, you'd
find the single binary version of Sourcegraph.

This didn't really catch on with the rest of the company/team, I left the
Delivery team so can't push this idea forward myself, and we ended up
adding many more app-specific behaviors in the codebase. If we wanted this
idea to come back, we would need to vet each `deploy.IsApp()` case to see
which ones should apply to a 'single binary enterprise' version of Sourcegraph.

It's just not worth it right now, so removing.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>