observation: set Tracer in TestContext
We recently set tracer in TestContextTB. We should do the same for
TestContext which is also used a fair amount. Tracer is set in prod, so
exercising the paths in tests is valuable.

Note: This is take two on this commit since I forgot to push my fixes
before merging :O

Test Plan: CI